Move utility methods on the base controller, update affected code.
